+++ docker/events.go
@@ -38,8 +38,13 @@ func EmitStatus(e *core.NormalizedEmitter, r io.Reader, options *core.PipelineOp
 
 		line, err := s.ProcessJSONMessage(&m)
 		if err != nil {
+			e.Emit(core.Logs, &core.LogsArgs{
+				Logs:   err.Error() + "\n",
+				Stream: "docker",
+			})
 			return err
 		}
+
 		e.Emit(core.Logs, &core.LogsArgs{
 			Logs:   line,
 			Stream: "docker",