+++ changelog.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- docker-build bugfix: if the dockerfile build fails then the error is displayed and the step fails (#424)
+
 ## v1.0.1264 (2015-06-18)
 
 - Collect report artifacts from a step even if it failed (#428)